More adjustments based on code review.
